create view

+++ ManageProject/src/main/java/vn/iotstar/Repository/LectureRepository.java
@@ -5,8 +5,7 @@ import org.springframework.stereotype.Repository;
 
 import vn.iotstar.Entity.Lecture;
 
-
 @Repository
-public interface LectureRepository extends JpaRepository<Lecture, Long>{
+public interface LectureRepository extends JpaRepository<Lecture, Integer>{
 
 }
